refactor code simulation into the file gam.c => simulation should contain only the functions mandatory for the main
